requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
binaryornot==0.4.4
bitarray==2.9.2
black==24.8.0
cached-property==1.5.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cssbeautifier==1.15.1
cytoolz==1.0.0
Django==5.1.1
django-browser-reload==1.15.0
django-htmx==1.19.0
djlint==1.35.2
EditorConfig==0.12.4
eth-account==0.13.4
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==1.2.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.10
isort==5.13.2
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
lru-dict==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==3.0.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.9
py-ecc==7.0.1
py-evm==0.10.1b2
py-geth==5.0.0
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.1
rlp==4.0.1
safe-pysha3==1.0.4
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.1
text-unidecode==1.3
toolz==1.0.0
tqdm==4.66.5
trie==3.0.1
types-python-dateutil==2.9.0.20241003
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urbitob==0.0.4
urllib3==2.2.3
web3==7.3.0
websockets==13.1
yarl==1.13.1
